+++ core/genrcpsr.h
@@ -65,7 +65,7 @@
 
    bool                           ParseDefgeneric(Environment *,const char *);
    bool                           ParseDefmethod(Environment *,const char *);
-   Defmethod                     *AddMethod(Environment *,Defgeneric *,Defmethod *,int,short,Expression *,
+   Defmethod                     *AddMethod(Environment *,Defgeneric *,Defmethod *,int,unsigned,Expression *,
                                             int,int,CLIPSLexeme *,Expression *,char *,bool);
    void                           PackRestrictionTypes(Environment *,RESTRICTION *,Expression *);
    void                           DeleteTempRestricts(Environment *,Expression *);
